+++ XpringKit/XRPClient.swift
@@ -6,8 +6,8 @@ public class XRPClient {
   ///
   /// - Parameters:
   ///   - grpcURL: A remote URL for a rippled gRPC service.
-  ///   - useNewProtocolBuffers:  If `true`, then the new protocol buffer implementation from rippled will be used. Defaults to false.
-  public init(grpcURL: String, useNewProtocolBuffers: Bool = false) {
+  ///   - useNewProtocolBuffers:  If `true`, then the new protocol buffer implementation from rippled will be used. Defaults to true.
+  public init(grpcURL: String, useNewProtocolBuffers: Bool = true) {
     let defaultClient: XRPClientDecorator = useNewProtocolBuffers ?
       DefaultXRPClient(grpcURL: grpcURL) :
       LegacyDefaultXRPClient(grpcURL: grpcURL)
@@ -23,13 +23,26 @@ public class XRPClient {
     return try decoratedClient.getBalance(for: address)
   }
 
-  /// Retrieve the transaction status for a given transaction hash.
+  /// Retrieve the transaction status for a Payment given transaction hash.
   ///
-  /// - Parameter transactionHash: The hash of the transaction.
-  /// - Throws: An error if there was a problem communicating with the XRP Ledger.
+  /// - Note: This method will only work for Payment type transactions which do not have the tf_partial_payment attribute set.
+  /// - SeeAlso: https://xrpl.org/payment.html#payment-flags
+  ///
+  /// - Parameter transactionHash The hash of the transaction.
   /// - Returns: The status of the given transaction.
   public func getTransactionStatus(for transactionHash: TransactionHash) throws -> TransactionStatus {
-    return try decoratedClient.getTransactionStatus(for: transactionHash)
+    return try self.paymentStatus(for: transactionHash)
+  }
+
+  /// Retrieve the transaction status for a Payment given transaction hash.
+  ///
+  /// - Note: This method will only work for Payment type transactions which do not have the tf_partial_payment attribute set.
+  /// - SeeAlso: https://xrpl.org/payment.html#payment-flags
+  ///
+  /// - Parameter transactionHash The hash of the transaction.
+  /// - Returns: The status of the given transaction.
+  public func paymentStatus(for transactionHash: TransactionHash) throws -> TransactionStatus {
+    return try self.decoratedClient.paymentStatus(for: transactionHash)
   }
 
   /// Send XRP to a recipient on the XRP Ledger.